digital-pathology-master_modified/src/SlideProcessing/requirements.txt
openslide-python==1.1.1
Pillow==5.3.0
